pyproject.toml
[tool.poetry]
name = "knowledgeable-cobra"
version = "0.1.0"
description = ""
authors = ["Zeyang Lin <4020306+linzeyang@users.noreply.github.com>"]
readme = "README.md"
license = "LICENSE.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = "^0.0.30"}
pydantic = {extras = ["email"], version = "<2"}
python-dotenv = "^1.0.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
motor = {extras = ["srv"], version = "^3.3.2"}
pymilvus = "^2.3.3"
qdrant-client = "^1.6.9"
weaviate-client = "^3.25.3"
cohere = "^4.36"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = "^0.0.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.coverage.run]
branch = true
source = [ "./" ]

[tool.coverage.report]
omit = [ "test_*.py", "conftest.py" ]

[tool.isort]
py_version = 310
profile = "black"

[tool.pylint.format]
max-line-length = "88"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [ "./" ]
python_files = [ "test_*.py" ]
console_output_style = "count"

[tool.ruff]
line-length = 88
target-version = "py310"
